UPSTREAM: thunderbolt: Allow disabling XDomain protocol
This allows disabling XDomain protocol completely if the user does not
plan to use the USB4/Thunderbolt peer-to-peer functionality, or for
security reasons.

XDomain protocol is enabled by default but with this commit it is
possible to disable it by passing "xdomain=0" as module parameter (or
through the kernel command line).

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Acked-by: Yehezkel Bernat <YehezkelShB@gmail.com>
(cherry picked from commit 5ca67688256a0aeede5cba288eaef4d8e4a9e622)

BUG=b:175139600
TEST=No TBT/USB4 regressions

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I5f620e0a984a43e184070a5a4e1cf7a48bd25b3f
Signed-off-by: Chintan Patel <chintan.m.patel@intel.corp-partner.google.com>
Signed-off-by: Daniel Campello <campello@chromium.org>