CHROMIUM: thunderbolt: Allow disabling XDomain protocol
This allows user to disable XDomain protocol completely, for example for
security reasons or if the user simply does not plan to use any of the
Thunderbolt/USB4 peer-to-peer functionality. XDomain is enabled by
default but with this commit it is possible to disable it by passing
"xdomain=0" as module parameter (or through the kernel command line).

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

BUG=b:170261534
TEST=Confirm P2P is disabled (no thunderbolt0 in ifconfig)

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I4f8fdae2edaacb66ac2dbf43f20bc01a6e5a588e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2547940
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>