CHROMIUM: arm64: dts: qcom: Properly setup lazor sc7180-lite skus
Some of the SKUs of sc7180-trogdor-lazor are the "lite" version of the
processor.

NOTE: We only started getting "lite" SKUs at at -rev3.  However, it's
not really worth it to add a special case for -rev1 and -rev2.  Why?
* The worst that happens here is that DDR will transition to a higher
  frequency a little sooner than it should on -rev1 and -rev2.
* -rev1 and -rev2 don't really need long term support, so having them
  run at a slightly-less optimal CPU frequency should be fine and is
  better than adding to the tangle of device trees.

NOTE ALSO: I'm still not 100% sure that we want to land support for
"lite" SKUs upstream using these types of dts tweaks, which is why it
makes sense to make this a CHROMIUM patch for now.  See b:171811744
for some details.

BUG=b:171811744
TEST=Build and boot

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I73dda5414969eb1d3b3f3ba1e11a6ecd8be1f94c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2503460
Reviewed-by: Stephen Boyd <swboyd@chromium.org>