BACKPORT: platform/x86: Intel PMT Crashlog capability driver
Add support for the Intel Platform Monitoring Technology crashlog
interface. This interface provides a few sysfs values to allow for
controlling the crashlog telemetry interface as well as a character
driver to allow for mapping the crashlog memory region so that it can be
accessed after a crashlog has been recorded.

This driver is meant to only support the server version of the crashlog
which is identified as crash_type 1 with a version of zero. Currently no
other types are supported.

Signed-off-by: Alexander Duyck <alexander.h.duyck@linux.intel.com>
Signed-off-by: David E. Box <david.e.box@linux.intel.com>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>

(cherry picked from commit 5ef9998c96b0c99c49c202054586967e609286d2)

Conflicts:
   drivers/platform/x86/Kconfig
   drivers/platform/x86/Makefile

Kconfig resolution was similar to the previous commit. Leave existing
content and only introduce the block for INTEL_PMT_CRASHLOG.

Makefile resolution was similar to the previous commit. Only introduce
the single line for intel_pmt_crashlog.o.

BUG=b:187383885
TEST=Verify kernel builds for volteer.
     `emerge-volteer chromeos-kernel-5_4`
TEST=Verify PMT sysfs exists on volteer.
     `ls /sys/class/intel_pmt`

Signed-off-by: Ed Baker <edward.baker@intel.com>
Change-Id: Id1144f266a735ce84636a2a0ba543964e79258ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2668186
Reviewed-by: Alex Levin <levinale@google.com>