CHROMIUM: drm/img-rogue: Don't overwrite devfreq->min/max_freq
In InitDVFS() the img-rogue driver assigns values to min/max_freq fields
of the devfreq device. This is not necessary, the fields don't limit or
specify the min/max frequency of the device, but represent the limits set
by userspace, the devfreq core initializes the fields with the min/max
values from the OPP table (see devfreq_add_device()). Moreover these
fields cease to exist in kernels >= v5.5, which use dev_pm_qos to
limit the frequency range.

BUG=b:163880423
TEST=pre-cq passes

Change-Id: I93692fa3e4e89d50ad8c8665e38afa67c0519673
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2514542
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>