CHROMIUM: drm: panel: simple: Add vrefresh for BOE NV110WTM-N61
Upstream has removed the vrefresh field in commit 0425662fdf05 ("drm:
Nuke mode->vrefresh").  That CL not only adds support to calculate the
vrefresh dynamically in some cases but it also removes all users which
makes it really hard to backport (because there are more users
upstream than downstream).  For now, let's add the vrefresh back in
for BOE NV110WTM-N61.  If later someone backports the CL removing
vrefresh it will be a super obvious compile-time error.

NOTE: if we skip this patch things seem to work for the most part, but
I'm a little worried that there's some bit of code somewhere that was
relying on it.  Best to take it.

BUG=b:169288627, b:171536503
TEST=Boot coachz

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ice6d4fb9488ec667056d9a627f52fcb9e9785527
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2522097
Reviewed-by: Rob Clark <robdclark@chromium.org>