CHROMIUM: thunderbolt: Add support for de-authorizing devices
In some cases it is useful to be able de-authorize devices. For example
if user logs out the userspace can have a policy that disconnects PCIe
devices until logged in again. This is only possible for software based
connection manager as it directly controls the tunnels.

For this reason make the authorized attribute allow writing 0 which
makes the software connection manager to tear down the corresponding
PCIe tunnel. In order to allow userspace to query whether writing 0 is
allowed expose deauthorize attribute under each domain. If this is 1 it
means de-authorization is supported, 0 otherwise.

While there correct tb_domain_approve_switch() kernel-doc and
description of authorized attribute to mention that it is only about
PCIe tunnels.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>

Conflicts:
	Documentation/ABI/testing/sysfs-bus-thunderbolt

BUG=b:172095861
TEST=echo 0 > /sys/bus/thunderbolt/X-Y/authorized
TEST=Confirm PCIe tunnels are deactivated

Signed-off-by: Casey Bowman <casey.g.bowman@intel.com>
Change-Id: I500c5048e55df1c900b200c3520c1672477ab4cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2531362
Reviewed-by: Rajat Jain <rajatja@google.com>
Commit-Queue: Rajat Jain <rajatja@google.com>
Tested-by: Rajat Jain <rajatja@google.com>