FROMGIT: net: ipa: ignore the microcontroller log event
The IPA-resident microcontroller has the ability to log various
activity in an area of IPA shared memory.  When the microcontroller
starts it generates an event to the AP to provide information about
the log.

We don't support reading this log, and we can safely ignore the
event.  So do that rather than treating the log info event we
receive as "unsupported."

Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
(cherry picked from commit 0a5096ec2a3501bd468853153a8129bb682e4f09
 https://git.kernel.org/pub/scm/linux/kernel/git/netdev/net-next.git master)

BUG=b:174161909
TEST=Booted trogdor, modem call functional, s/r work

Signed-off-by: Sujit Kautkar <sujitka@chromium.org>
Change-Id: I6554b2ca15d965783806f5b2f370be5a86e2257d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2559834
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Evan Green <evgreen@chromium.org>