UPSTREAM: drm/fourcc: add table describing AMD modifiers bit layout
The table describes how each bit in the u64 value is used. Explicitly
state which values a field can take if we have defines for them. Also
add a note when a field isn't always populated.

Forcing people to update the table when changing the bit layout should
make it more obvious when there's a mistake, I hope.

If we get to the point where the bit layout gets more complicated, it
might be worth it to split the table into multiple tables (e.g. one for
GFX8, one for GFX9+, and so on).

Signed-off-by: Simon Ser <contact@emersion.fr>
Cc: Bas Nieuwenhuizen <bas@basnieuwenhuizen.nl>
Cc: Alex Deucher <alexdeucher@gmail.com>
Cc: Daniel Vetter <daniel@ffwll.ch>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(cherry picked from commit 544645f2ec1af910284ebde00da2a6cfab7cc8c1)
Signed-off-by: Bas Nieuwenhuizen <basni@chromium.org>

BUG=b:184949073
TEST=compile chromeos-kernel-5_4 for Zork

Change-Id: I0f1bbffb60400fdef16f524d7ba46c12935e97f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2818084
Tested-by: Bas Nieuwenhuizen <basni@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Queue: Bas Nieuwenhuizen <basni@chromium.org>