UPSTREAM: drm/msm/shrinker: Only iterate dontneed objs
In situations where the GPU is mostly idle, all or nearly all buffer
objects will be in the inactive list.  But if the system is under memory
pressure (from something other than GPU), we could still get a lot of
shrinker calls.  Which results in traversing a list of thousands of objs
and in the end finding nothing to shrink.  Which isn't so efficient.

Instead split the inactive_list into two lists, one inactive objs which
are shrinkable, and a second one for those that are not.  This way we
can avoid traversing objs which we know are not shrinker candidates.

v2: Fix inverted logic think-o

Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 3edfa30f2340e6c361b34fc0c53a5f3d3bbf9704)

BUG=b:184092409
TEST=boot with mem=4g, open a bunch of web pages (separate windows) and
     run mmm_donut, and confirm at least mouse is still responsive

Change-Id: I8b7a84b145c5f86d2716654e00511576fe25f86b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2793041
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Rob Clark <robdclark@chromium.org>
Auto-Submit: Rob Clark <robdclark@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>