FIXUP: CHROMIUM: Update core scheduling CPU selection to upstream
The following patch was rolled into the upstream patch, but we have
to backport it to downstream kernels: https://lkml.org/lkml/2020/11/17/2137
  "sched/fair: Fix pick_task_fair crashes due to empty rbtree"

pick_next_entity() is passed curr == NULL during core-scheduling. Due to
this, if the rbtree is empty, the 'left' variable is set to NULL within
the function. This can cause crashes within the function.

This is not an issue if put_prev_task() is invoked on the currently
running task before calling pick_next_entity(). However, in core
scheduling, it is possible that a sibling CPU picks for another RQ in
the core, via pick_task_fair(). This remote sibling would not get any
opportunities to do a put_prev_task().

Fix it by refactoring pick_task_fair() such that pick_next_entity() is
called with the cfs_rq->curr. This will prevent pick_next_entity() from
crashing if its rbtree is empty.

Also this fixes another possible bug where update_curr() would not be
called on the cfs_rq hierarchy if the rbtree is empty. This could effect
cross-cpu comparison of vruntime.

BUG=b:221071901
TEST=build and CQ tests

Suggested-by: Vineeth Remanan Pillai <viremana@linux.microsoft.com>
Signed-off-by: Peter Zijlstra <peterz@infradead.org>
Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>
Change-Id: I46513d52821fb3c8cf2f65b88df9d1dbfa94c0e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3495503
Tested-by: Joel Fernandes <joelaf@google.com>
Reviewed-by: Vineeth Pillai <vineethrp@google.com>
Commit-Queue: Joel Fernandes <joelaf@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>