FROMLIST: venus: Limit HFI sessions to the maximum supported
Currently we rely on firmware to return error when we reach the maximum
supported number of sessions. But this errors are happened at reqbuf
time which is a bit later. The more reasonable way looks like is to
return the error on driver open.

To achieve that modify hfi_session_create to return error when we reach
maximum count of sessions and thus refuse open.

Signed-off-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
(am from https://lore.kernel.org/patchwork/patch/1342027/)

BUG=b:171358961
TEST=tast run ${DUT} video.*

Signed-off-by: Fritz Koenig <frkoenig@chromium.org>
Change-Id: I765670c016aebdcc7d5173c99989884ac81717db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2558488
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Alexandre Courbot <acourbot@chromium.org>