BACKPORT: usb: typec: Add type sysfs attribute file for partners
USB Power Delivery Specification defines a set of product
types for partners and cables. The product type can be read
from the ID Header VDO which is the first object in the
response to the Discover Identity command. This attribute
will display the product type of the partner. The cables
already have the attribute.

This sysfs attribute file is only created for the partners
and cables if the product type is really known in the
driver. Some interfaces do not give access to the Discover
Identity response from the partner or cable, but they may
still supply the product type separately in some cases.

When the product type of the partner or cable is detected,
uevent is also raised with PRODUCT_TYPE set to show the
actual product type (for example PRODUCT_TYPE=host).

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Link: https://lore.kernel.org/r/20201126115735.50529-1-heikki.krogerus@linux.intel.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit ab37fa851c488be805f6568ecaabb67b13cd937c)
[bleung: Switched back to sprintf instead of sysfs_emit]
Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=b:179157034
TEST=Check that partners and cables have the new "type" attribute.

Change-Id: I975edbb545cf29ac9b48bb92f435bc5183411aeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2668033
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Benson Leung <bleung@google.com>
Tested-by: Benson Leung <bleung@google.com>